+++ wiki/Changelog.md
@@ -12,6 +12,7 @@ Text display:
   * Option WideIndic for double-width display of wide Indic characters (#553) (experimental).
   * Not combining characters to unexisting glyphs.
   * Preventing mangled digit rendering if surrounded by certain scripts (~#285).
+  * Preventing mangled combining characters by drawing them separately (~#553, #295).
 
 Window handling:
   * Tweaked DPI changes to stabilize roundtrips of font and window size (#470).